backend/requirements.txt
flask
flask-cors
python-dotenv
requests
weasyprint
gunicorn
google-api-python-client
google-auth-oauthlib
google-auth-httplib2apscheduler
